requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pdfkit==1.0.0
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
